Let the user switch projects from the pick neighbourhood screen
